refactor/cleanup basically everything graphql-related
