+++ lib/server/graphql/payment-processor-adapters/braintree.js
@@ -1,9 +1,7 @@
 'use strict';
 
-const constants = require('../../constants');
+const constants = require('../../../constants');
 const BRAINTREE = constants.PAYMENT_PROCESSORS.BRAINTREE;
-
-let models;
 
 const braintreeAdapter = {
   add: function(){
@@ -16,9 +14,4 @@ const braintreeAdapter = {
   }
 };
 
-const braintreeAdapterFactory = (instanceModels)=> {
-  models = instanceModels;
-  return braintreeAdapter;
-};
-
-module.exports = braintreeAdapterFactory;
+module.exports = braintreeAdapter;
