+++ lib/server/graphql/payment-processor-adapters/stripe.js
@@ -3,11 +3,9 @@
 // TODO: pull secret key from env var
 // NB: this is not a real secret key so u dudes who think ur clever can suck it!
 const stripe = require('stripe')('sk_test_W6L09JRZ1YR4Ua0KuDCDTST3');
-const constants = require('../../constants');
+const constants = require('../../../constants');
 const STRIPE = constants.PAYMENT_PROCESSORS.STRIPE;
 const STRIPE_PLAN_ID = constants.STRIPE_PLAN_ID;
-
-let models;
 
 /**
  * checkCustomerStatus - check if current user has a payment processor
@@ -161,9 +159,4 @@ const stripeAdapter = {
   }
 };
 
-const stripeAdapterFactory = (instanceModels)=> {
-  models = instanceModels;
-  return stripeAdapter;
-};
-
-module.exports = stripeAdapterFactory;
+module.exports = stripeAdapter;
