+++ lib/server/graphql/types/credit.js
@@ -1,0 +1,15 @@
+'use strict';
+
+const graphql = require('graphql');
+
+const CreditType = new graphql.GraphQLObjectType({
+  name: 'Credit',
+  fields: {
+    id: {type: graphql.GraphQLString},
+    amount: {type: graphql.GraphQLInt},
+    created: {type: graphql.GraphQLString},
+    type: {type: graphql.GraphQLString}
+  }
+});
+
+module.exports = CreditType;
