+++ lib/server/graphql/types/debit.js
@@ -1,0 +1,15 @@
+'use strict';
+
+const graphql = require('graphql');
+
+const DebitType = new graphql.GraphQLObjectType({
+  name: 'Debit',
+  fields: {
+    id: {type: graphql.GraphQLString},
+    amount: {type: graphql.GraphQLInt},
+    created: {type: graphql.GraphQLString},
+    type: {type: graphql.GraphQLString}
+  }
+});
+
+module.exports = DebitType;
