+++ lib/server/routes/graphql.js
@@ -6,7 +6,6 @@ const inherits = require('util').inherits;
 const graphqlHTTP = require('express-graphql');
 const graphql = require('graphql');
 const authenticate = require('../middleware').authenticate;
-const schema = require('../../storage/graphql/schema');
 
 /**
  * GraphQLRouter
@@ -23,8 +22,14 @@ function GraphQLRouter(options) {
     return new GraphQLRouter(options);
   }
 
+  const storage = options.storage;
+
   Router.apply(this, arguments);
 
+  this.schema = storage.graphqlSchema;
+  this.graphqlHTTPMiddleware = graphqlHTTP({
+    schema: this.schema
+  }, options);
   this._verify = authenticate(this.storage);
 }
 
@@ -33,16 +38,14 @@ function GraphQLRouter(options) {
  */
 inherits(GraphQLRouter, Router);
 
-const options = {
-  mutation: true,
-  allowMongoIDMutation: false
-};
+GraphQLRouter.prototype.processRequest = function(req, res, next) {
+  this.schema.middleware(req, res, (err) => {
+    if (err) {
+      throw err;
+    }
 
-GraphQLRouter.prototype.graphqlMiddleware = function(req, res, next){
-  graphqlHTTP(
-      {schema: schema(this.storage.models, req)},
-      options
-  )(req, res, next);
+    this.graphqlHTTPMiddleware(req, res, next);
+  });
 };
 
 
@@ -53,7 +56,7 @@ GraphQLRouter.DEFAULTS = {
 
 GraphQLRouter.prototype._definitions = function() {
   return [
-    ['POST', '/graphql', this._verify, this.graphqlMiddleware]
+    ['POST', '/graphql', this._verify, this.processRequest]
   ];
 };
