+++ lib/storage/index.js
@@ -21,6 +21,7 @@ function Storage(options) {
   this._options = options;
   this.connection = this._connect();
   this.models = this._createBoundModels();
+  this.graphqlSchema = this._createBoundGraphqlSchema(this.models);
 
   this.connection.on('error', function(err) {
     log.error('failed to open database connection, reason: %s', err.message);
@@ -107,6 +108,12 @@ Storage.prototype._createBoundModels = function() {
   }
 
   return bound;
+};
+
+Storage.prototype._createBoundGraphqlSchema = function(models){
+  const graphql = require('../server/graphql');
+
+  return graphql.bindSchema(this.models);
 };
 
 module.exports = Storage;
